# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bug Tracker API",
        "description": "Dokumentasi REST API Bug Tracker untuk QA, DEV, dan PM.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/bugs/{bugId}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List komentar untuk 1 bug",
                "description": "Mengambil semua komentar yang terkait dengan bug tertentu.",
                "operationId": "a12fbaef7a7e74c87db37ab4c78b8a4e",
                "parameters": [
                    {
                        "name": "bugId",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar komentar berhasil diambil"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Tambah komentar ke bug",
                "description": "Menambahkan komentar baru pada bug tertentu.",
                "operationId": "24046f3b2d00ab0060b1b7068f2a28ee",
                "parameters": [
                    {
                        "name": "bugId",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Sudah di-test, masih muncul error di step 3."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Komentar berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Hapus komentar",
                "description": "Menghapus komentar (hanya oleh pemilik komentar atau role tertentu).",
                "operationId": "b683c30a239aa780ffc7e9b219e90811",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID komentar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Komentar berhasil dihapus"
                    },
                    "404": {
                        "description": "Komentar tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/bugs/summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Ringkasan bug per periode",
                "description": "Mengembalikan total bug dan breakdown status untuk rentang tanggal tertentu.",
                "operationId": "7f63d0e3042f061bb4e6bd5aa444b296",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Tanggal awal (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Tanggal akhir (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2025-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil ringkasan bug"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/bugs/by-status": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Jumlah bug per status",
                "description": "Mengembalikan jumlah bug untuk setiap status (OPEN, IN_PROGRESS, RESOLVED, CLOSED).",
                "operationId": "9683cfdc71320ae9dc2979ae4882515f",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/bugs/by-severity": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Jumlah bug per severity",
                "description": "Mengembalikan jumlah bug untuk setiap severity (LOW, MEDIUM, HIGH, CRITICAL).",
                "operationId": "f853011a2fe440abcb535ce3d41958b7",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari /api/login atau /api/register.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}